Control.Monad.Logger.CallStack

Description

Log functions using CallStack support in place of Template Haskell

Since: 0.3.19

Documentation

class ToLogStr msg where

Types that can be converted to a LogStr. Instances for types from the text library use a UTF-8 encoding. Instances for numerical types use a decimal encoding.

Methods

toLogStr :: msg -> LogStr

data LogStr

Log message builder. Use (<>) to append two LogStr in O(1).

data Loc

Constructors

Loc

Fields

loc_filename :: String

loc_package :: String

loc_module :: String

loc_start :: CharPos

loc_end :: CharPos

class Monad m => MonadLogger m where

A Monad which has the ability to log messages in some manner.

Minimal complete definition

Nothing

Methods

monadLoggerLog :: ToLogStr msg => Loc -> LogSource -> LogLevel -> msg -> m ()

default monadLoggerLog :: (MonadLogger m', MonadTrans t, MonadLogger (t m'), ToLogStr msg, m ~ t m') => Loc -> LogSource -> LogLevel -> msg -> m ()

class (MonadLogger m, MonadIO m) => MonadLoggerIO m where

An extension of MonadLogger for the common case where the logging action is a simple IO action. The advantage of using this typeclass is that the logging function itself can be extracted as a first-class value, which can make it easier to manipulate monad transformer stacks, as an example.

Since: 0.3.10

Minimal complete definition

Nothing

Methods

askLoggerIO :: m (Loc -> LogSource -> LogLevel -> LogStr -> IO ())

Request the logging function itself.

Since: 0.3.10

default askLoggerIO :: (MonadTrans t, MonadLoggerIO n, m ~ t n) => m (Loc -> LogSource -> LogLevel -> LogStr -> IO ())

data LogLevel

Constructors

LevelDebug

LevelInfo

LevelWarn

LevelError

LevelOther Text

type LogLine = (Loc, LogSource, LogLevel, LogStr)

Since: 0.3.32

type LogSource = Text

newtype LoggingT m a

Monad transformer that adds a new logging function.

Since: 0.2.2

Constructors

LoggingT

Fields

runLoggingT :: (Loc -> LogSource -> LogLevel -> LogStr -> IO ()) -> m a

newtype NoLoggingT m a

Monad transformer that disables logging.

Since: 0.2.4

Constructors

NoLoggingT

Fields

runNoLoggingT :: m a

newtype WriterLoggingT m a

Since: 0.3.28

Constructors

WriterLoggingT

Fields

unWriterLoggingT :: m (a, DList LogLine)

fromLogStr :: LogStr -> ByteString

Converting LogStr to ByteString.

runStderrLoggingT :: MonadIO m => LoggingT m a -> m a

Run a block using a MonadLogger instance which prints to stderr.

Since: 0.2.2

runStdoutLoggingT :: MonadIO m => LoggingT m a -> m a

Run a block using a MonadLogger instance which prints to stdout.

Since: 0.2.2

runChanLoggingT :: MonadIO m => Chan LogLine -> LoggingT m a -> m a

Run a block using a MonadLogger instance which writes tuples to an unbounded channel.

The tuples can be extracted (ie. in another thread) with unChanLoggingT or a custom extraction funtion, and written to a destination.

Since: 0.3.17

runFileLoggingT :: MonadBaseControl IO m => FilePath -> LoggingT m a -> m a

Run a block using a MonadLogger instance which appends to the specified file.

Since: 0.3.22

unChanLoggingT :: (MonadLogger m, MonadIO m) => Chan LogLine -> m void

Read logging tuples from an unbounded channel and log them into a MonadLoggerIO monad, forever.

For use in a dedicated thread with a channel fed by runChanLoggingT.

Since: 0.3.17

withChannelLogger

Arguments

:: (MonadBaseControl IO m, MonadIO m)

=> Int
Number of messages to keep

-> LoggingT m a

-> LoggingT m a

Within the LoggingT monad, capture all log messages to a bounded channel of the indicated size, and only actually log them if there is an exception.

Since: 0.3.2

filterLogger :: (LogSource -> LogLevel -> Bool) -> LoggingT m a -> LoggingT m a

Only log messages passing the given predicate function.

This can be a convenient way, for example, to ignore debug level messages.

Since: 0.3.13

mapNoLoggingT :: (m a -> n b) -> NoLoggingT m a -> NoLoggingT n b

Map the unwrapped computation using the given function.

Since: 0.3.29

execWriterLoggingT :: Functor m => WriterLoggingT m a -> m [LogLine]

Run a block using a MonadLogger instance. Return logs in a list | @since 0.3.28

runWriterLoggingT :: Functor m => WriterLoggingT m a -> m (a, [LogLine])

Run a block using a MonadLogger instance. Return a value and logs in a list | @since 0.3.28

mapLoggingT :: (m a -> n b) -> LoggingT m a -> LoggingT n b

Map the unwrapped computation using the given function.

Since: 0.3.29

logDebugS :: Q Exp

Generates a function that takes a LogSource and Text and logs a LevelDebug message. Usage:

$logDebugS "SomeSource" "This is a debug log message"

logInfoS :: Q Exp

See logDebugS

logWarnS :: Q Exp

See logDebugS

logErrorS :: Q Exp

See logDebugS

logOtherS :: Q Exp

Generates a function that takes a LogSource, a level name and a Text and logs a LevelOther message. Usage:

$logOtherS "SomeSource" "My new level" "This is a log message"

liftLoc :: Loc -> Q Exp

Lift a location into an Exp.

Since: 0.3.1

logDebugN :: MonadLogger m => Text -> m ()

logInfoN :: MonadLogger m => Text -> m ()

logWarnN :: MonadLogger m => Text -> m ()

logErrorN :: MonadLogger m => Text -> m ()

logOtherN :: MonadLogger m => LogLevel -> Text -> m ()

logWithoutLoc :: (MonadLogger m, ToLogStr msg) => LogSource -> LogLevel -> msg -> m ()

Since: 0.3.23

logDebugNS :: MonadLogger m => LogSource -> Text -> m ()

logInfoNS :: MonadLogger m => LogSource -> Text -> m ()

logWarnNS :: MonadLogger m => LogSource -> Text -> m ()

logErrorNS :: MonadLogger m => LogSource -> Text -> m ()

logOtherNS :: MonadLogger m => LogSource -> LogLevel -> Text -> m ()

logDebugCS :: MonadLogger m => CallStack -> Text -> m ()

Logs a message with location given by CallStack. See CallStack for more convenient functions for CallStack based logging.

Since: 0.3.19

logInfoCS :: MonadLogger m => CallStack -> Text -> m ()

See logDebugCS

Since: 0.3.19

logWarnCS :: MonadLogger m => CallStack -> Text -> m ()

See logDebugCS

Since: 0.3.19

logErrorCS :: MonadLogger m => CallStack -> Text -> m ()

See logDebugCS

Since: 0.3.19

logOtherCS :: MonadLogger m => CallStack -> LogLevel -> Text -> m ()

See logDebugCS

Since: 0.3.19

defaultLogStr :: Loc -> LogSource -> LogLevel -> LogStr -> LogStr

defaultLoc :: Loc

dummy location, used with logWithoutLoc

Since: 0.3.23

defaultOutput :: Handle -> Loc -> LogSource -> LogLevel -> LogStr -> IO ()

A default implementation of monadLoggerLog that accepts a file handle as the first argument.

This is used in the definition of runStdoutLoggingT:

runStdoutLoggingT :: MonadIO m => LoggingT m a -> m a
runStdoutLoggingT action =
    runLoggingT action (defaultOutput stdout)

Since: 0.3.36

logDebug :: (HasCallStack, MonadLogger m) => Text -> m ()

Logs a message with the location provided by an implicit CallStack.

Since: 0.3.19

logInfo :: (HasCallStack, MonadLogger m) => Text -> m ()

See logDebug

Since: 0.3.19

logWarn :: (HasCallStack, MonadLogger m) => Text -> m ()

See logDebug

Since: 0.3.19

logError :: (HasCallStack, MonadLogger m) => Text -> m ()

See logDebug

Since: 0.3.19

logOther :: (HasCallStack, MonadLogger m) => LogLevel -> Text -> m ()

See logDebug

Since: 0.3.25

logDebugSH :: (HasCallStack, MonadLogger m, Show a) => a -> m ()

Logs a showable value with the location provided by an implicit CallStack.

Since: 0.3.25

logInfoSH :: (HasCallStack, MonadLogger m, Show a) => a -> m ()

See logDebugSH

Since: 0.3.25

logWarnSH :: (HasCallStack, MonadLogger m, Show a) => a -> m ()

See logDebugSH

Since: 0.3.25

logErrorSH :: (HasCallStack, MonadLogger m, Show a) => a -> m ()

See logDebugSH

Since: 0.3.25

logOtherSH :: (HasCallStack, MonadLogger m, Show a) => LogLevel -> a -> m ()

See logDebugSH

Since: 0.3.25