Control.Monad.Logger

Description

This module provides the facilities needed for a decoupled logging system.

The MonadLogger class is implemented by monads that give access to a logging facility. If you're defining a custom monad, then you may define an instance of MonadLogger that routes the log messages to the appropriate place (e.g., that's what yesod-core's GHandler does). Otherwise, you may use the LoggingT monad included in this module (see runStderrLoggingT). To simply discard log message, use NoLoggingT.

As a user of the logging facility, we provide you some convenient Template Haskell splices that use the MonadLogger class. They will record their source file and position, which is very helpful when debugging. See logDebug for more information.

MonadLogger

class Monad m => MonadLogger m where

Methods

monadLoggerLog :: ToLogStr msg => Loc -> LogSource -> LogLevel -> msg -> m ()

data LogLevel

Constructors

LevelDebug

LevelInfo

LevelWarn

LevelError

LevelOther Text

type LogSource = Text

Helper transformer

newtype LoggingT m a

Monad transformer that adds a new logging function.

Since 0.2.2

Constructors

LoggingT

Fields

runLoggingT :: (Loc -> LogSource -> LogLevel -> LogStr -> IO ()) -> m a

runStderrLoggingT :: MonadIO m => LoggingT m a -> m a

Run a block using a MonadLogger instance which prints to stderr.

Since 0.2.2

runStdoutLoggingT :: MonadIO m => LoggingT m a -> m a

Run a block using a MonadLogger instance which prints to stdout.

Since 0.2.2

withChannelLogger

Arguments

:: (MonadBaseControl IO m, MonadIO m)

=> Int
Number of mesasges to keep

-> LoggingT m a

-> LoggingT m a

Within the LoggingT monad, capture all log messages to a bounded channel of the indicated size, and only actually log them if there is an exception.

Since 0.3.2

newtype NoLoggingT m a

Monad transformer that disables logging.

Since 0.2.4

Constructors

NoLoggingT

Fields

runNoLoggingT :: m a

TH logging

logDebug :: Q Exp

Generates a function that takes a Text and logs a LevelDebug message. Usage:

 $(logDebug) "This is a debug log message"

logInfo :: Q Exp

See logDebug

logWarn :: Q Exp

See logDebug

logError :: Q Exp

See logDebug

logOther :: Text -> Q Exp

Generates a function that takes a Text and logs a LevelOther message. Usage:

 $(logOther "My new level") "This is a log message"

TH logging with source

logDebugS :: Q Exp

Generates a function that takes a LogSource and Text and logs a LevelDebug message. Usage:

 $logDebugS "SomeSource" "This is a debug log message"

logInfoS :: Q Exp

See logDebugS

logWarnS :: Q Exp

See logDebugS

logErrorS :: Q Exp

See logDebugS

logOtherS :: Q Exp

Generates a function that takes a LogSource, a level name and a Text and logs a LevelOther message. Usage:

 $logOtherS "SomeSource" "My new level" "This is a log message"

TH util

liftLoc :: Loc -> Q Exp

Lift a location into an Exp.

Since 0.3.1

Non-TH logging

logDebugN :: MonadLogger m => Text -> m ()

logInfoN :: MonadLogger m => Text -> m ()

logWarnN :: MonadLogger m => Text -> m ()

logErrorN :: MonadLogger m => Text -> m ()

logOtherN :: MonadLogger m => LogLevel -> Text -> m ()

Non-TH logging with source

logDebugNS :: MonadLogger m => Text -> Text -> m ()

logInfoNS :: MonadLogger m => Text -> Text -> m ()

logWarnNS :: MonadLogger m => Text -> Text -> m ()

logErrorNS :: MonadLogger m => Text -> Text -> m ()

logOtherNS :: MonadLogger m => Text -> LogLevel -> Text -> m ()