Control.Monad.Logger.CallStack

Description

Log functions using CallStack support in place of Template Haskell

Since: 0.3.19

Documentation

logDebug :: (HasCallStack, MonadLogger m) => Text -> m ()

Logs a message with the location provided by an implicit CallStack.

Since: 0.3.19

logInfo :: (HasCallStack, MonadLogger m) => Text -> m ()

See logDebug

Since: 0.3.19

logWarn :: (HasCallStack, MonadLogger m) => Text -> m ()

See logDebug

Since: 0.3.19

logError :: (HasCallStack, MonadLogger m) => Text -> m ()

See logDebug

Since: 0.3.19

logOther :: (HasCallStack, MonadLogger m) => LogLevel -> Text -> m ()

See logDebug

Since: 0.3.25

logDebugSH :: (HasCallStack, MonadLogger m, Show a) => a -> m ()

Logs a showable value with the location provided by an implicit CallStack.

Since: 0.3.25

logInfoSH :: (HasCallStack, MonadLogger m, Show a) => a -> m ()

See logDebugSH

Since: 0.3.25

logWarnSH :: (HasCallStack, MonadLogger m, Show a) => a -> m ()

See logDebugSH

Since: 0.3.25

logErrorSH :: (HasCallStack, MonadLogger m, Show a) => a -> m ()

See logDebugSH

Since: 0.3.25

logOtherSH :: (HasCallStack, MonadLogger m, Show a) => LogLevel -> a -> m ()

See logDebugSH

Since: 0.3.25